simulation/requirements.txt
azure-iot-device
azure-iot-hub
azure-storage-queue
asyncio
